requirements.txt
meson
ninja
torch
pyyaml
numpy
colorama